+++ pyophidia/utils/cwl2oph.py
@@ -25,7 +25,6 @@ sys.path.insert(0, os.path.dirname(previous_dir))
 sys.path.insert(0, "..")
 from pyophidia import Workflow, Experiment, Task
 
-
 def run():
     print("Parse arguments", file=sys.stderr)
     parser = argparse.ArgumentParser()
@@ -351,7 +350,6 @@ def run():
         )
     print("Add task '" + args.name + "'", file=sys.stderr)
     print(repr(e1))
-
 
 if __name__ == "__main__":
     run()