+++ pyophidia/workflow.py
@@ -57,6 +57,7 @@ def _dependency_check(dependency):
         raise AttributeError("Dependency must be prov, cwl or graphviz")
 
 
+
 class Workflow:
     """
     Submits, cancels and monitors a workflow experiment execution (a workflow)